Squashed 'src/bamdb/' changes from 5a20dae..1692669
1692669 Merge pull request #17 from D-Lo/fix
350073f Flag off more code when building in ro mode
e18ea97 Merge pull request #16 from D-Lo/refactor
77c7a49 Allow for read only builds that do not require ck
ca6e880 Renaming files for more clarity

git-subtree-dir: src/bamdb
git-subtree-split: 16926694801c6e6bba0c195f92e736e2a6707589

+++ src/bam_api.c
@@ -389,7 +389,7 @@ void print_sequence_row(bam_sequence_row_t *row) {
         printf("f:%g", *(float *)aux->val);
         break;
       case 'd':
-        printf("d:%g", *(float *)aux->val);
+        printf("d:%g", *(double *)aux->val);
         break;
       case 'Z':
       case 'H':
